requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
django-countries==7.3.2
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysql-connector-python==8.0.29
mysqlclient==2.1.0
Pillow==9.1.1
protobuf==4.21.0
psycopg2==2.9.3
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
